+++ src/common/constants.ts
@@ -1,0 +1,6 @@
+export const DHFS_SCHEME = 'dhfs';
+// const CONNECT_COMMAND = "dh-in-vscode.connect";
+export const RUN_CODE_COMMAND = 'dh-in-vscode.runCode';
+export const RUN_SELECTION_COMMAND = 'dh-in-vscode.runSelection';
+export const SELECT_CONNECTION_COMMAND = 'dh-in-vscode.selectConnection';
+export const SELECTED_CONNECTION_STORAGE_KEY = 'selectedConnection';
