+++ src/common/types.d.ts
@@ -1,0 +1,1 @@
+export type ConnectionType = 'DHC' | 'DHE';
